increased token lifetime to 18 months as using implicit grant
This follows from the GOV.UK documentation which states a token can be refreshed for up to 18 months: https://developer.service.hmrc.gov.uk/api-documentation/docs/authorisation/user-restricted-endpoints

As this is using implicit grant, we will allow the token to live for 18 months, however, we will forget the access token on the frontend every day at midnight, forcing the user to authenticate again to get an access token.

However, the use will not have to re-authorize the app for 18 months.